added constans folder and file for CATEGORY_BUTTONS that includes category 0 - all recipes
